nextjs update to 12.2.5 + mods to work
