+++ lib/tests/cors.test.js
@@ -1,3 +1,7 @@
+/**
+ * @jest-environment node
+ */
+
 import { createMocks } from "node-mocks-http";
 import { cors } from "@lib/middleware";
