+++ lib/tests/jsonValidator.test.js
@@ -1,3 +1,7 @@
+/**
+ * @jest-environment node
+ */
+
 import { createMocks } from "node-mocks-http";
 import { jsonValidator } from "@lib/middleware";
 import templatesSchema from "../middleware/schemas/templates.schema.json";
