+++ lib/tests/middleware.test.js
@@ -1,3 +1,7 @@
+/**
+ * @jest-environment node
+ */
+
 import { middleware } from "@lib/middleware";
 import { createMocks } from "node-mocks-http";
