+++ lib/tests/sessionExists.test.js
@@ -1,3 +1,7 @@
+/**
+ * @jest-environment node
+ */
+
 import { createMocks } from "node-mocks-http";
 import { unstable_getServerSession } from "next-auth/next";
 import { sessionExists } from "@lib/middleware";
