+++ package.json
@@ -56,7 +56,7 @@
     "lodash.set": "^4.3.2",
     "markdown-to-jsx": "^7.1.7",
     "mmmagic": "^0.5.3",
-    "next": "^12.1.4",
+    "next": "^12.2.5",
     "next-auth": "^4.10.3",
     "next-i18next": "^11.0.0",
     "notifications-node-client": "^5.1.1",
@@ -84,8 +84,9 @@
     "@storybook/manager-webpack5": "^6.4.20",
     "@storybook/react": "^6.4.20",
     "@storybook/storybook-deployer": "^2.8.11",
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
+    "@testing-library/dom": "^8.17.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/react-hooks": "^8.0.1",
     "@testing-library/user-event": "^14.1.1",
     "@types/classnames": "^2.2.11",
@@ -108,7 +109,6 @@
     "autoprefixer": "^10.4.4",
     "aws-sdk-client-mock": "^0.6.2",
     "axe-core": "^4.4.1",
-    "babel-jest": "^27.5.1",
     "babel-loader": "^8.2.4",
     "babel-plugin-dynamic-import-node": "^2.3.3",
     "babel-plugin-transform-require-context": "^0.1.1",
@@ -127,8 +127,9 @@
     "eslint-plugin-testing-library": "^5.2.1",
     "husky": "^7.0.4",
     "ioredis-mock": "^8.2.2",
-    "jest": "^27.5.1",
+    "jest": "28",
     "jest-css-modules": "^2.1.0",
+    "jest-environment-jsdom": "^29.0.1",
     "jest-fetch-mock": "^3.0.3",
     "jest-mock-extended": "^2.0.5",
     "node-mocks-http": "^1.11.0",
@@ -140,9 +141,9 @@
     "sass": "^1.49.11",
     "sass-loader": "^12.6.0",
     "tailwindcss": "^3.0.23",
-    "ts-jest": "^27.1.4",
+    "ts-jest": "^28.0.8",
     "ts-node": "^10.9.1",
-    "typescript": "^4.6.3",
+    "typescript": "^4.8.2",
     "webpack": "^5.71.0"
   },
   "browserslist": [
